requirements.txt
flask
git-python
